+++ hw05-jdbc-hard/src/main/java/ru/otus/hw/repositories/GenreRowMapper.java
@@ -1,0 +1,19 @@
+package ru.otus.hw.repositories;
+
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Component;
+import ru.otus.hw.models.Genre;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@Component
+public class GenreRowMapper implements RowMapper<Genre> {
+
+    @Override
+    public Genre mapRow(ResultSet rs, int i) throws SQLException {
+        long id = rs.getLong("id");
+        String name = rs.getString("name");
+        return new Genre(id, name);
+    }
+}
