+++ hw05-jdbc-hard/src/main/java/ru/otus/hw/repositories/JdbcBookRepository.java
@@ -1,8 +1,6 @@
 package ru.otus.hw.repositories;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
@@ -11,15 +9,11 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import ru.otus.hw.exceptions.EntityNotFoundException;
-import ru.otus.hw.models.Author;
 import ru.otus.hw.models.Book;
-import ru.otus.hw.models.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -31,7 +25,7 @@ public class JdbcBookRepository implements BookRepository {
 
     private final NamedParameterJdbcOperations jdbcOperations;
 
-    private final GenreRepository genreRepository;
+    private final BookResultSetExtractor bookResultSetExtractor;
 
     @Override
     public Optional<Book> findById(long id) {
@@ -51,7 +45,7 @@ public class JdbcBookRepository implements BookRepository {
                 WHERE
                     book.id = :id
                 """;
-        List<Book> books = jdbcOperations.query(sql, Collections.singletonMap("id", id), new BookResultSetExtractor());
+        List<Book> books = jdbcOperations.query(sql, Collections.singletonMap("id", id), bookResultSetExtractor);
         if (Objects.isNull(books) || books.isEmpty()) {
             return Optional.empty();
         }
@@ -74,7 +68,7 @@ public class JdbcBookRepository implements BookRepository {
                     INNER JOIN books_genres AS books_genres ON book.id = books_genres.book_id
                     INNER JOIN genres AS genre ON books_genres.genre_id = genre.id
                 """;
-        return jdbcOperations.query(sql, new BookResultSetExtractor());
+        return jdbcOperations.query(sql, bookResultSetExtractor);
     }
 
     @Override
@@ -145,40 +139,6 @@ public class JdbcBookRepository implements BookRepository {
     private void removeGenresRelationsFor(Book book) {
         String sql = "DELETE FROM books_genres WHERE book_id = :book_id";
         jdbcOperations.update(sql, Collections.singletonMap("book_id", book.getId()));
-    }
-
-    @RequiredArgsConstructor
-    private static class BookResultSetExtractor implements ResultSetExtractor<List<Book>> {
-
-        @Override
-        public List<Book> extractData(ResultSet rs) throws SQLException, DataAccessException {
-            Map<Long, Book> books = new HashMap<>();
-            Map<Long, Author> authors = new HashMap<>();
-            Map<Long, Genre> genres = new HashMap<>();
-            while (rs.next()) {
-                Long bookId = rs.getLong("book_id");
-                Long genreId = rs.getLong("genre_id");
-                Long authorId = rs.getLong("author_id");
-                Author author = authors.get(authorId);
-                if (Objects.isNull(author)) {
-                    author = new Author(authorId, rs.getString("author_full_name"));
-                    authors.put(authorId, author);
-                }
-                Genre genre = genres.get(genreId);
-                if (Objects.isNull(genre)) {
-                    genre = new Genre(genreId, rs.getString("genre_name"));
-                    genres.put(genreId, genre);
-                }
-                Book book = books.get(bookId);
-                if (Objects.isNull(book)) {
-                    book = new Book(bookId, rs.getString("book_title"), author, new ArrayList<>());
-                    book.setGenres(new ArrayList<>());
-                    books.put(bookId, book);
-                }
-                book.getGenres().add(genre);
-            }
-            return new ArrayList<>(books.values());
-        }
     }
 
     private static class BookGenreRelationRowMapper implements RowMapper<BookGenreRelation> {
